4-assistant-function-calling/requirements.txt
colorama==0.4.4
python-dotenv==1.0.0
openai==1.3.5
streamlit==1.28.1
tavily-python==0.2.8
streamlit-chat==0.1.1
tavily-python==0.2.8
